Update snort
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@26414 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 8ae1fde0404bf7a46a65147aa3271c90abc7d074

+++ src/router/snort/src/preprocessors/HttpInspect/include/hi_ui_iis_unicode_map.h
@@ -1,6 +1,7 @@
 /****************************************************************************
  *
- * Copyright (C) 2003-2011 Sourcefire, Inc.
+ * Copyright (C) 2014 Cisco and/or its affiliates. All rights reserved.
+ * Copyright (C) 2003-2013 Sourcefire, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License Version 2 as
@@ -15,7 +16,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
  ****************************************************************************/
  
@@ -38,7 +39,7 @@
 */
 #define HI_UI_NON_ASCII_CODEPOINT -1
 
-int hi_ui_parse_iis_unicode_map(int **iis_unicode_map, char *filename,
+int hi_ui_parse_iis_unicode_map(uint8_t **iis_unicode_map, char *filename,
                                 int iCodePage);
 
 #endif
